% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THREC.R
\docType{data}
\name{Treeht}
\alias{Treeht}
\title{Tree data from the long-term forest experiments in Sweden}
\format{
\subsection{\code{Treeht}}{

A data frame with 334 rows and 9 columns:
\describe{
\item{yta}{Plot number}
\item{rev}{Revision number}
\item{t_l}{Species code}
\item{hojd}{Sampled tree height in meters}
\item{dia}{Diameter at breast height in centimeter}
\item{hdom}{Height of the tree with the largest diameter per plot regardless of the species}
\item{ddom}{Largest diameter per plot regardless of the species}
\item{QMD}{Quadratic mean diameter in centimeter per plot}
\item{G}{Basal area in meter square per hectare}
}
}
}
\source{
\url{doi:10.1016/j.foreco.2023.120843}
}
\usage{
data(Treeht)
}
\description{
A subset of the data used for the development of multilevel mixed-effect height functions for the Swedish long-term forest experiments
Report ...
}
\examples{
data(Treeht)

}
\keyword{datasets}
